refactor: change file structure to be consistent with routes
